refactor: NavigationBar 위치 수정
[#804]
