+++ frontend/src/components/ui/Navigator/NavigationBar/BasePanel.tsx
