+++ frontend/src/components/ui/Navigator/NavigationBar/CloseButton.tsx
