+++ frontend/src/components/ui/Navigator/NavigationBar/LastPanel.tsx
