+++ frontend/src/components/ui/Navigator/NavigationBar/hooks/useNavigationBar.ts
